+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.14.5] - 2024-03-20
+
+## Added
+
+- Support for audio_events in Batch CLI.
+- Support `types` whitelist for audio events.
+
 ## [1.14.4] - 2024-03-04
 
 ## Added